fix(data): align mergeProperties dedup with persisted-JSON equivalence (#196)
Address Codex review: the merged result is stored via JSON.stringify, so
distinguishing NaN/null/undefined in the in-memory dedup key only produced
a persisted [null, null] duplicate that the storage layer collapses anyway.

Key the list dedupe by the same equivalence the storage layer uses for
no-op detection (JSON.stringify of the key-sorted canonical form). This
still fixes the real bug — key-order-sensitivity that left reordered-equal
objects (e.g. {id,alias} vs {alias,id}) as duplicates — while staying
consistent with persistence: NaN/undefined collapse to null exactly as
properties_json serialization does, so a merge never keeps two items the
DB would persist as identical.

Extract stableJsonValue/jsonValuesEqual into internals/jsonCanonical.ts as
the single source of truth, shared by txEngine and mergeProperties so the
two equivalences can't drift. Update the #196 reproducer tests accordingly.
